refactor and adds database queries for tile info instead of export
